Unit Testing over Voting Station complete
I moved Driver_Mainframe to unitTests, it seemed more logical. I used a
StationsControllerStub and not the Driver_StationsController so unless
someone else need it it can be deleted. One of the tests veryfies that
after 2 minutes one can't revote so for that the test really waits 2
minutes, it is normal. About this I changed the canVote time checking
from 20 sec to 120 sec, maybe the 20 sec was on purpose but I think it
is better to put directly what was in the requirements of the project or
we risk forgetting to change it in the end.