full practice test environment
Change PracticeStation to make calls deterministic.
Improve generic test to handle waiting (in choosing+guide)

to do: create every test in the test plan that I did not already cover

by the why: timeGuard is necessery to assure that timeout is identify.
right know, it work when timeGuard < 400 millis, so no to much trouble.