Fixed bug in dictionaries
You probably haven't noticed (because you don't do anything) but the
hebrew dictionary had a serious problem. Git didn't copy the source file
correctly (it all came out gibbrish). Now we read the dictionaries from
files in the dictionaries folder on system start (that way there are no
copy errors as the files are copied in the correct UTF-8 format)