Mainframe & VotingStation integration - final
No evidence of any problem in mainframe crashes (everything goes as
expected). That means the bug in backup happens because of the actual
saving to files and not because of flawed synchronization of threads (I
told you I was perfect). Emil, now it's up to you. Do your magic, find
end eliminate this bug!